requirements.txt
absl-py==2.1.0
addict==2.4.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.1.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coacd==1.0.5
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
decorator==5.2.0
Deprecated==1.2.18
dill==0.3.9
etils==1.12.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.56.0
freetype-py==2.5.1
fsspec==2024.6.1
genesis-world==0.2.1
glfw==2.8.0
grpcio==1.70.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
libigl==2.5.1
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==2.1.2
mpmath==1.3.0
mujoco==3.2.5
mypy-extensions==1.0.0
narwhals==1.27.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
open3d==0.19.0
opencv-python==4.11.0.86
OpenEXR==3.3.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.0
pooch==1.8.2
proglog==0.1.10
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycollada==0.9
pydantic==2.7.1
pydantic_core==2.18.2
PyGEL3D==0.5.2
pyglet==2.1.2
pygltflib==1.16.0
Pygments==2.19.1
pymeshlab==2023.12.post2
PyOpenGL==3.1.9
pyparsing==3.2.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyvista==0.44.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.23.1
-e git+https://github.com/leggedrobotics/rsl_rl@6b9057008cf9e85d4306cd8a4914a751392e8892#egg=rsl_rl
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
scooby==0.10.0
screeninfo==0.8.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
taichi==1.7.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tetgen==0.6.4
threadpoolctl==3.5.0
tifffile==2025.2.18
tomli==2.2.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vtk==9.3.1
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
zipp==3.21.0
